fix(clustering): fix data corruption during centroid reordering
Fixed a critical bug where centroids were being overwritten in-place during the sorting process.
The previous implementation used a shallow copy (auto tmp = centroidsData), causing the original data to be modified before all values were correctly remapped. Replaced it with a deep copy (std::vector) to preserve original values during the reordering loop, aligning the behavior with the TypeScript implementation.

+++ src/writers/lod_writer.cpp
@@ -250,7 +250,7 @@ void writeLod(const std::string& filename, const DataTable* dataTable, DataTable
     if (!mNode.lods.empty()) {
       j["lods"] = json::object();
       for (auto const& [lodKey, lodVal] : mNode.lods) {
-        j["lods"][std::to_string(lodKey)] = {{"file", lodVal.file}, {"offset", lodVal.offset}, {"count", lodVal.count}};
+        j["lods"][std::to_string(static_cast<int>(lodKey))] = {{"file", lodVal.file}, {"offset", lodVal.offset}, {"count", lodVal.count}};
       }
     }
     return j;
